Move the view components to their own files
