+++ LumiformApp/LumiformApp/QuestionView.swift
@@ -1,0 +1,28 @@
+//
+//  QuestionView.swift
+//  LumiformApp
+//
+//  Created by Aly Yakan on 25/07/2025.
+//
+
+import SwiftUI
+import Lumiform
+
+struct QuestionView: View {
+    let question: Question
+
+    var body: some View {
+        switch question {
+        case .text(let textQuestion):
+            Text(textQuestion.title)
+                .font(.subheadline)
+        case .image(let imageQuestion):
+            // TODO: Display image
+            Text(imageQuestion.title)
+                .font(.subheadline)
+        @unknown default:
+            Text("Unknown question format")
+                .font(.subheadline)
+        }
+    }
+}
